Algorithm specifications refactoring.
Moved the algorithm specifications object into separate module.
Moved the algorithm specs converters to the dtos module.
The 'algorithm_specifications' cannot be inside the algorithms module, due to multiple imports.
Ticket in jira: MIP-65
