+++ mipengine/controller/api/algorithm_specifications_dtos.py
@@ -1,6 +1,4 @@
 from abc import ABC
-from enum import Enum
-from enum import unique
 from typing import Any
 from typing import Dict
 from typing import List
@@ -8,41 +6,21 @@ from typing import Optional
 
 from pydantic import BaseModel
 
+from mipengine.algorithm_specification import AlgorithmSpecification
+from mipengine.algorithm_specification import InputDataSpecification
+from mipengine.algorithm_specification import InputDataSpecifications
+from mipengine.algorithm_specification import InputDataStatType
+from mipengine.algorithm_specification import InputDataType
+from mipengine.algorithm_specification import ParameterEnumSpecification
+from mipengine.algorithm_specification import ParameterEnumType
+from mipengine.algorithm_specification import ParameterSpecification
+from mipengine.algorithm_specification import ParameterType
+from mipengine.controller import algorithms_specifications
+
 
 class ImmutableBaseModel(BaseModel, ABC):
     class Config:
         allow_mutation = False
-
-
-@unique
-class InputDataType(Enum):
-    REAL = "real"
-    INT = "int"
-    TEXT = "text"
-    JSONOBJECT = "jsonObject"
-
-
-@unique
-class InputDataStatType(Enum):
-    NUMERICAL = "numerical"
-    NOMINAL = "nominal"
-
-
-@unique
-class ParameterType(str, Enum):
-    REAL = "real"
-    INT = "int"
-    TEXT = "text"
-    BOOLEAN = "boolean"
-    DICT = "dict"
-
-
-@unique
-class ParameterEnumType(str, Enum):
-    LIST = "list"
-    INPUT_VAR_CDE_ENUMS = "input_var_CDE_enums"
-    FIXED_VAR_CDE_ENUMS = "fixed_var_CDE_enums"
-    INPUT_VAR_NAMES = "input_var_names"
 
 
 class InputDataSpecificationDTO(ImmutableBaseModel):
@@ -92,3 +70,132 @@ class AlgorithmSpecificationDTO(ImmutableBaseModel):
 
 class AlgorithmSpecificationsDTO(ImmutableBaseModel):
     __root__: List[AlgorithmSpecificationDTO]
+
+
+def _convert_inputdata_specification_to_dto(self: InputDataSpecification):
+    # The only difference of the DTO is that it's stattypes is Optional,
+    # due to the fact that the datasets/data_model variables are added.
+    return InputDataSpecificationDTO(
+        label=self.label,
+        desc=self.desc,
+        types=self.types,
+        stattypes=self.stattypes,
+        notblank=self.notblank,
+        multiple=self.multiple,
+        enumslen=self.enumslen,
+    )
+
+
+def _get_data_model_input_data_specification_DTO():
+    return InputDataSpecificationDTO(
+        label="Data model of the data.",
+        desc="The data model that the algorithm will run on.",
+        types=[InputDataType.TEXT],
+        notblank=True,
+        multiple=False,
+        stattypes=None,
+        enumslen=None,
+    )
+
+
+def _get_datasets_input_data_specification_DTO():
+    return InputDataSpecificationDTO(
+        label="Set of data to use.",
+        desc="The set of data to run the algorithm on.",
+        types=[InputDataType.TEXT],
+        notblank=True,
+        multiple=True,
+        stattypes=None,
+        enumslen=None,
+    )
+
+
+def _get_filters_input_data_specification_DTO():
+    return InputDataSpecificationDTO(
+        label="filter on the data.",
+        desc="Features used in my algorithm.",
+        types=[InputDataType.JSONOBJECT],
+        notblank=False,
+        multiple=False,
+        stattypes=None,
+        enumslen=None,
+    )
+
+
+def _convert_inputdata_specifications_to_dto(spec: InputDataSpecifications):
+    # In the DTO the datasets, data_model and filter parameters are added from the engine.
+    # These parameters are not added by the algorithm developer.
+    y = _convert_inputdata_specification_to_dto(spec.y)
+    x = _convert_inputdata_specification_to_dto(spec.x) if spec.x else None
+    return InputDataSpecificationsDTO(
+        y=y,
+        x=x,
+        data_model=_get_data_model_input_data_specification_DTO(),
+        datasets=_get_datasets_input_data_specification_DTO(),
+        filter=_get_filters_input_data_specification_DTO(),
+    )
+
+
+def _convert_parameter_enum_specification_to_dto(spec: ParameterEnumSpecification):
+    return ParameterEnumSpecificationDTO(
+        type=spec.type,
+        source=spec.source,
+    )
+
+
+def _convert_parameter_specification_to_dto(spec: ParameterSpecification):
+    return ParameterSpecificationDTO(
+        label=spec.label,
+        desc=spec.desc,
+        types=spec.types,
+        notblank=spec.notblank,
+        multiple=spec.multiple,
+        default=spec.default,
+        enums=(
+            _convert_parameter_enum_specification_to_dto(spec.enums)
+            if spec.enums
+            else None
+        ),
+        dict_keys_enums=(
+            _convert_parameter_enum_specification_to_dto(spec.dict_keys_enums)
+            if spec.dict_keys_enums
+            else None
+        ),
+        dict_values_enums=(
+            _convert_parameter_enum_specification_to_dto(spec.dict_values_enums)
+            if spec.dict_values_enums
+            else None
+        ),
+        min=spec.min,
+        max=spec.max,
+    )
+
+
+def _convert_algorithm_specification_to_dto(spec: AlgorithmSpecification):
+    # Converting to a DTO does not include the flags.
+    return AlgorithmSpecificationDTO(
+        name=spec.name,
+        desc=spec.desc,
+        label=spec.label,
+        inputdata=_convert_inputdata_specifications_to_dto(spec.inputdata),
+        parameters=(
+            {
+                name: _convert_parameter_specification_to_dto(value)
+                for name, value in spec.parameters.items()
+            }
+            if spec.parameters
+            else None
+        ),
+    )
+
+
+def _get_algorithm_specifications_dtos() -> AlgorithmSpecificationsDTO:
+    return AlgorithmSpecificationsDTO(
+        __root__=[
+            _convert_algorithm_specification_to_dto(algorithm)
+            for algorithm in algorithms_specifications.values()
+        ]
+    )
+
+
+algorithm_specifications_dtos = _get_algorithm_specifications_dtos()
