+++ mipengine/controller/api/endpoint.py
@@ -5,8 +5,10 @@ from quart import Blueprint
 from quart import request
 
 from mipengine.controller import config as controller_config
-from mipengine.controller.algorithm_specifications import algorithm_specifications
 from mipengine.controller.api.algorithm_request_dto import AlgorithmRequestDTO
+from mipengine.controller.api.algorithm_specifications_dtos import (
+    algorithm_specifications_dtos,
+)
 from mipengine.controller.api.loggers import loggers
 from mipengine.controller.api.validator import BadRequest
 from mipengine.controller.cleaner import Cleaner
@@ -51,7 +53,7 @@ async def get_data_models_attributes() -> dict:
 
 @algorithms.route("/algorithms", methods=["GET"])
 async def get_algorithms() -> str:
-    return algorithm_specifications.get_enabled_algorithm_dtos().json()
+    return algorithm_specifications_dtos.json()
 
 
 @algorithms.route("/algorithms/<algorithm_name>", methods=["POST"])
